LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/container-name-parsing-expected.txt


PASS e.style['container-name'] = "initial" should set the property value
PASS e.style['container-name'] = "inherit" should set the property value
PASS e.style['container-name'] = "unset" should set the property value
PASS e.style['container-name'] = "revert" should set the property value
PASS e.style['container-name'] = "none" should set the property value
PASS e.style['container-name'] = "foo" should set the property value
PASS e.style['container-name'] = "BAR" should set the property value
PASS e.style['container-name'] = "foo bar" should set the property value
PASS e.style['container-name'] = "none none" should not set the property value
PASS e.style['container-name'] = "foo, bar" should not set the property value
PASS e.style['container-name'] = "#fff" should not set the property value
PASS e.style['container-name'] = "1px" should not set the property value
PASS e.style['container-name'] = "default" should not set the property value
FAIL e.style['container-name'] = "\"initial\"" should not set the property value assert_equals: expected "" but got "\"initial\""
FAIL e.style['container-name'] = "\"inherit\"" should not set the property value assert_equals: expected "" but got "\"inherit\""
FAIL e.style['container-name'] = "\"unset\"" should not set the property value assert_equals: expected "" but got "\"unset\""
FAIL e.style['container-name'] = "\"revert\"" should not set the property value assert_equals: expected "" but got "\"revert\""
FAIL e.style['container-name'] = "\"none\"" should not set the property value assert_equals: expected "" but got "\"none\""
FAIL e.style['container-name'] = "\"foo\"" should not set the property value assert_equals: expected "" but got "\"foo\""
